feat(US-11): Refactor file structuring for client folder
